apps/backend/requirements.txt
alembic==1.13.1
aniso8601==9.0.1
black==24.2.0
blinker==1.8.2
click==8.1.7
flake8==7.0.0
Flask==3.0.2
flask-restx==1.3.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.6
MarkupSafe==3.0.2
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pyflakes==3.2.0
pytest==8.1.1
python-dotenv==1.0.1
pytz==2024.2
six==1.16.0
SQLAlchemy==2.0.28
typing_extensions==4.12.2
Werkzeug==3.0.6